Nevermind, need the arbitrary precision

+++ FrameworkHs/Prims.hs
@@ -43,9 +43,9 @@ import Data.Int
 --------------------------------------------------------------------------------
 -- Terminal Type Definitions ---------------------------------------------------
 
-data UVar = UV String Int64 deriving (Read, Show, Eq)
-data FVar = FV Int64        deriving (Read, Show, Eq)
-data Label = L String Int64 deriving (Read, Show, Eq)
+data UVar = UV String Integer deriving (Read, Show, Eq)
+data FVar = FV Integer        deriving (Read, Show, Eq)
+data Label = L String Integer deriving (Read, Show, Eq)
 
 data Reg = RAX | RCX | RDX | RBX | RBP | RSI | RDI | R8 | R9
          | R10 | R11 | R12 | R13 | R14 | R15
@@ -54,5 +54,5 @@ data Reg = RAX | RCX | RDX | RBX | RBP | RSI | RDI | R8 | R9
 data Relop = LT | LTE | EQ | GT | GTE               deriving (Read,Show,Eq,Ord)
 data Binop = MUL | ADD | SUB | LOGAND | LOGOR | SRA deriving (Read,Show,Eq,Ord)
 
-data Disp = Disp Reg Int64
+data Disp = Disp Reg Integer
 data Ind = Ind Reg Reg
